- added repoName and repoPass as config params - modified some statements to simplify or adapt to our use
having 2 databases per environment, with the original authors code, the snapshots would be stored to the same restic repository. That's why we added another config parameter in repoName. The default repoPassword would always be the env name, which is not very secure either, so we made the password configurable as well